requirements.txt
addict==2.4.0
antlr4-python3-runtime==4.9.3
asttokens==2.4.1
attrs==23.2.0
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
chumpy==0.70
click==8.1.7
cmake==3.29.2
comm==0.2.2
ConfigArgParse==1.7
contourpy==1.2.1
cycler==0.12.1
dash==2.16.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dataclasses==0.6
decorator==5.1.1
exceptiongroup==1.2.1
executing==2.0.1
fastjsonschema==2.19.1
filelock==3.13.4
Flask==3.0.3
fonttools==4.51.0
freetype-py==2.4.0
fvcore==0.1.5.post20221221
idna==3.7
imageio==2.34.1
importlib_metadata==7.1.0
install==1.3.5
iopath==0.1.10
ipython==8.23.0
ipywidgets==8.1.2
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.4.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_core==5.7.2
jupyterlab_widgets==3.0.10
kiwisolver==1.4.5
lit==18.1.4
loguru==0.7.2
MarkupSafe==2.1.5
matplotlib==3.8.4
matplotlib-inline==0.1.7
mpmath==1.3.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
numpy==1.23.1
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cccl-cu11==11.8.89
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.2.10.91
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.4.91
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu11==2.14.3
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu11==11.7.91
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
open3d==0.18.0
opencv-python==4.9.0.80
orjson==3.10.3
packaging==24.0
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.0
plotly==5.21.0
plyfile==1.0.3
portalocker==2.8.2
prompt-toolkit==3.0.43
psbody-mesh==0.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyglet==1.5.27
Pygments==2.17.2
PyOpenGL==3.1.0
pyparsing==3.1.2
pyquaternion==0.9.9
pyrender==0.1.45
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.2
referencing==0.34.0
requests==2.31.0
retrying==1.3.4
rpds-py==0.18.0
scikit-learn==1.4.2
scipy==1.13.0
shapely==2.0.4
six==1.16.0
smplx==0.1.28
stack-data==0.6.3
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
termcolor==2.4.0
threadpoolctl==3.4.0
torch==1.13.1
torch_trust_ncg==0.1.0
torchaudio==0.13.1
torchvision==0.14.1
tqdm==4.66.2
traitlets==5.14.3
trimesh==4.3.1
triton==2.2.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
wcwidth==0.2.13
Werkzeug==3.0.2
widgetsnbextension==4.0.10
yacs==0.1.8
zipp==3.18.1
